Frontend/requirements.txt
pysqlite3-binary
chromadb
PyPDF2
python-dotenv
fastapi
uvicorn
streamlit